Refactored and cleaned
