--- src/app/app-state.module.ts
@@ -1,7 +1,0 @@
-import { NgModule } from '@angular/core';
-import { NgxsDataPluginModule } from '@ngxs-labs/data';
-
-@NgModule({
-  exports: [NgxsDataPluginModule]
-})
-export class AppStateModule {}
