+++ src/app/core/models/ontology-node.ts
