+++ src/app/core/store/search/search.state.ts
@@ -2,7 +2,18 @@ import { Injectable } from '@angular/core';
 import { action, NgxsDataRepository, StateRepository } from '@ngxs-labs/data';
 import { State } from '@ngxs/store';
 
-import { SearchStateModel } from '../../models/search-state.model';
+import { OntologyNode } from '../../models/ontology-node';
+
+
+/**
+ * Search state model.
+ */
+export interface SearchStateModel {
+  /**
+   * The currently selected anatomical location.
+   */
+  location: OntologyNode | undefined;
+}
 
 /**
  * Contains the currently active search parameters.
@@ -16,9 +27,8 @@ import { SearchStateModel } from '../../models/search-state.model';
 })
 @Injectable()
 export class SearchState extends NgxsDataRepository<SearchStateModel>  {
-
   @action()
   setLocation(ontology: SearchStateModel): void {
-    this.ctx.setState((state) => ontology);
+    this.ctx.setState(ontology);
   }
 }
