+++ src/app/modules/ontology-exploration/ontology-search/ontology-search.component.html
@@ -1,10 +1,11 @@
 <form class="ccf-ontology-search" [ngClass]="{'autocomplete-open': autoCompleteOpen}">
   <mat-form-field appearance="outline" class="field">
-    <input class="input" type="text"  placeholder="Search ontology terms ..." matInput [formControl]="formControl"
-     [matAutocomplete]="autocomplete">
+    <input class="input" type="text" placeholder="Search ontology terms ..." matInput [formControl]="formControl"
+           [matAutocomplete]="autocomplete">
 
     <mat-autocomplete class="ccf-ontology-search detached autocomplete" [displayWith]="displayFormatter"
-                      (optionSelected)="onSelect($event)" (opened)="autoCompleteOpen = true" (closed)="autoCompleteOpen = false" #autocomplete>
+                      (optionSelected)="onSelect($event)" (opened)="autoCompleteOpen = true"
+                      (closed)="autoCompleteOpen = false" #autocomplete>
       <mat-option *ngFor="let option of filteredResults$ | async" [value]="option">
         <span class="prefix">{{option.displayLabel[0]}}</span>
         <span class="search-term">{{option.displayLabel[1]}}</span>
