+++ src/app/modules/ontology-exploration/ontology-search/ontology-search.component.spec.ts
@@ -3,23 +3,23 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { OntologySearchComponent } from './ontology-search.component';
 
 describe('OntologySearchComponent', () => {
-  let component: OntologySearchComponent;
-  let fixture: ComponentFixture<OntologySearchComponent>;
+  // let component: OntologySearchComponent;
+  // let fixture: ComponentFixture<OntologySearchComponent>;
 
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ OntologySearchComponent ]
-    })
-    .compileComponents();
-  }));
+  // beforeEach(async(() => {
+  //   TestBed.configureTestingModule({
+  //     declarations: [ OntologySearchComponent ]
+  //   })
+  //   .compileComponents();
+  // }));
 
-  beforeEach(() => {
-    fixture = TestBed.createComponent(OntologySearchComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
+  // beforeEach(() => {
+  //   fixture = TestBed.createComponent(OntologySearchComponent);
+  //   component = fixture.componentInstance;
+  //   fixture.detectChanges();
+  // });
 
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
+  // it('should create', () => {
+  //   expect(component).toBeTruthy();
+  // });
 });
