+++ src/app/modules/ontology-exploration/ontology-search/ontology-search.component.theme.scss
@@ -17,7 +17,7 @@
       .mat-form-field-outline {
         color: mat-color($foreground, divider);
       }
-  
+
       .mat-form-field-outline-thick {
         color: mat-color($foreground, divider);
       }
