+++ src/app/modules/ontology-exploration/ontology-search/ontology-search.module.ts
@@ -4,21 +4,22 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MatAutocompleteModule } from '@angular/material/autocomplete';
 import { MatIconModule } from '@angular/material/icon';
 import { MatInputModule } from '@angular/material/input';
-import { AppStateModule } from 'src/app/app-state.module';
 
 import { OntologySearchComponent } from './ontology-search.component';
+import { MatFormFieldModule } from '@angular/material/form-field';
+
 
 @NgModule({
-  declarations: [OntologySearchComponent],
   imports: [
     CommonModule,
     FormsModule,
-    MatAutocompleteModule,
-    MatIconModule,
-    MatInputModule,
     ReactiveFormsModule,
-    AppStateModule
+    MatAutocompleteModule,
+    MatFormFieldModule,
+    MatInputModule,
+    MatIconModule
   ],
+  declarations: [OntologySearchComponent],
   exports: [OntologySearchComponent]
 })
 export class OntologySearchModule {}
