+++ src/app/modules/ontology-exploration/ontology-selection/ontology-selection.module.ts
@@ -1,15 +1,16 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 
-import { OntologySelectionComponent } from './ontology-selection.component';
 import { OntologySearchModule } from '../ontology-search/ontology-search.module';
 import { OntologyTreeModule } from '../ontology-tree/ontology-tree.module';
+import { OntologySelectionComponent } from './ontology-selection.component';
+
 
 @NgModule({
   imports: [
-  CommonModule,
-  OntologySearchModule,
-  OntologyTreeModule
+    CommonModule,
+    OntologySearchModule,
+    OntologyTreeModule
   ],
   declarations: [OntologySelectionComponent],
   exports: [OntologySelectionComponent],
