--- src/app/shared/models/search-state.model.ts
@@ -1,11 +1,0 @@
-import { OntologyNode } from './ontology-node.model';
-
-/**
- * Search state model.
- */
-export interface SearchStateModel {
-  /**
-   * The currently selected anatomical location.
-   */
-  location: OntologyNode | undefined;
-}
