--- src/app/shared/services/flat-node/flat-node.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { FlatNode } from './flat-node.service';
-
-describe('FlatNodeService', () => {
-  let service: FlatNode;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(FlatNode);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
