--- src/app/shared/services/ontology-search/ontology-search.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { OntologySearchService } from './ontology-search.service';
-
-describe('OntologySearchService', () => {
-  let service: OntologySearchService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(OntologySearchService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
