+++ src/themes/_default.scss
@@ -1,7 +1,7 @@
 @use '~@angular/material/theming' as *;
 @use 'common/utility';
 @use '../app/core/header/header.component.theme' as header;
-@use '../app/modules/ontology-exploration/ontology-search/ontology-search.component.theme' as search;
+
 
 $primary-colors: utility.add-colors($mat-gray, (
   "white": white,
